доработано отображение даты/времени и фильтров
