+++ src/mock/filter.js
@@ -1,0 +1,7 @@
+const filtersName = [
+  `Everything`,
+  `Future`,
+  `Past`
+];
+
+export {filtersName};
