+++ src/view/event-edit.js
@@ -1,4 +1,4 @@
-import {createItems, dateFormat} from "../utils/utils.js";
+import {createItems, convertDate} from "../utils/utils.js";
 
 const getEventOffer = (offer, favorite) => {
   return (
@@ -49,12 +49,12 @@ const createEventEditTemplate = (events) => {
               <label class="visually-hidden" for="event-start-time-1">
                 From
               </label>
-              <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${dateFormat(startDate)}">
+              <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${convertDate(startDate)}">
               —
               <label class="visually-hidden" for="event-end-time-1">
                 To
               </label>
-              <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${dateFormat(endDate)}">
+              <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${convertDate(endDate)}">
             </div>
             <div class="event__field-group  event__field-group--price">
               <label class="event__label" for="event-price-1">
@@ -83,7 +83,9 @@ const createEventEditTemplate = (events) => {
             </section>
             <section class="event__section  event__section--destination">
               <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-              <p class="event__destination-description">${description}.</p>
+              ${description.length > 0 ? `
+              <p class="event__destination-description">${description}</p>` : ``
+    }
               ${photos.length > 0 ? `
               <div class="event__photos-container">
                 <div class="event__photos-tape">
